DataBased/requirements.txt
python-dotenv
google-cloud-aiplatform
langchain
langchain-community
langchain-core
langchain_google_genai
sentence-transformers
torch
langchain_chroma
langchain_google_vertexai
google_crc32c